feat(placeholder): expand custom placeholders and integrate MiniPlaceholders (#4)
- centralize placeholder processing into a single pipeline (remove scattered handling)
- add parameterized unit placeholders: {days: suffix} {hours: suffix} {minutes: suffix} {seconds: suffix}
  - hide token when value is 0 if suffix is provided
- add formatted remaining placeholder: {remaining:<pattern>} (e.g. {remaining:hh:mm:ss}, {remaining:d天 hh:mm:ss})
  - suppress zero-valued units when a unit suffix is present
- enable {animation:<id>} frames to pass through the same placeholder + MiniMessage pipeline
- integrate MiniPlaceholders as an optional dependency via TagResolver (no reflection)
  - render per-player where applicable (bossbar/title) for audience-aware placeholders
- update config comments and wiki/docs to describe new placeholder formats

BREAKING CHANGE: placeholder rendering has been refactored into a unified pipeline; any custom code that depended on old scattered render paths must be updated.

+++ src/main/java/top/ourisland/invertotimer/config/model/AnimationConfig.java
@@ -51,7 +51,7 @@ public record AnimationConfig(
                 if (!(o instanceof Map<?, ?> fm)) continue;
 
                 double durSec = YamlUtil.getDouble(fm, "duration", 1.0);
-                long durMs = Math.max(1, (long) Math.round(durSec * 1000.0));
+                long durMs = Math.max(1, Math.round(durSec * 1000.0));
                 String text = YamlUtil.getString(fm, "text", "");
 
                 frames.add(new Frame(durMs, text));
